refactor: Moved desktop Navbar in seperate folder, introduced new Navbar.tsx file to call both the Navbars depending on situation
