+++ app/[lang]/chapters/[slug]/layout.tsx
@@ -1,4 +1,4 @@
-import Navbar, { NavbarMobile } from 'components/Navbar'
+import Navbar from 'components/Navbar'
 import { chapters } from 'content'
 
 export default function Layout({ children, params }) {
@@ -6,13 +6,8 @@ export default function Layout({ children, params }) {
 
   return (
     <div className="ch1-background flex flex-col">
-      <div className="fix-grow-issue flex min-h-screen flex-col bg-back/75 drop-shadow-3xl backdrop-blur-4xl">
-        <div className="md:hidden">
-          <NavbarMobile params={params} />
-        </div>
-        <div className="hidden md:block">
-          <Navbar params={params} />
-        </div>
+      <div className="fix-grow-issue flex min-h-screen flex-col overflow-hidden bg-back/75 drop-shadow-3xl backdrop-blur-4xl">
+        <Navbar params={params} />
         {children}
       </div>
     </div>
