+++ components/Navbar/index.ts
@@ -1,3 +1,2 @@
 export { default } from './Navbar'
 export { default as Address } from './Address'
-export { default as NavbarMobile } from './NavbarMobile/NavbarMobile'
